requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
authlib==0.15.2           # via -r requirements.in
awswrangler==1.10.0       # via -r requirements.in
boto3==1.16.22            # via awswrangler
botocore==1.19.22         # via awswrangler, boto3, s3transfer
certifi==2019.11.28       # via requests
cffi==1.14.3              # via cryptography
chardet==3.0.4            # via requests
cryptography==3.2.1       # via authlib
hvac==0.10.5              # via -r requirements.in
idna==2.8                 # via requests
jmespath==0.9.4           # via boto3, botocore
numpy==1.19.4             # via -r requirements.in, awswrangler, pandas, pyarrow
packaging==20.4           # via sqlalchemy-redshift
pandas==1.1.4             # via -r requirements.in, awswrangler
psycopg2-binary==2.8.6    # via -r requirements.in, awswrangler
pyarrow==2.0.0            # via awswrangler
pycparser==2.20           # via cffi
pymysql==0.10.1           # via awswrangler
pyparsing==2.4.7          # via packaging
python-dateutil==2.8.1    # via botocore, pandas
pytz==2020.4              # via pandas
requests==2.22.0          # via -r requirements.in, hvac
s3transfer==0.3.1         # via boto3
six==1.14.0               # via cryptography, hvac, packaging, python-dateutil
sqlalchemy-redshift==0.8.1  # via awswrangler
sqlalchemy==1.3.20        # via awswrangler, sqlalchemy-redshift
urllib3==1.25.7           # via botocore, requests














